refactor: change the structure of card description
