+++ src/widgets/CardDescription/CardDescription.css
