+++ src/widgets/CardDescription/CardDetail.tsx
@@ -1,0 +1,7 @@
+import { CardDetailProps } from '../../types';
+
+export const CardDetail = (props: CardDetailProps): JSX.Element => (
+  <div>
+    <strong>{props.label}:</strong> {props.value}
+  </div>
+);
